% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{generate_pareto}
\alias{generate_pareto}
\title{Generating data from a Pareto Distribution.}
\usage{
generate_pareto(sample_size, shape, scale)
}
\arguments{
\item{sample_size}{number of observations}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\value{
Vector of Pareto distributed data of sample size \code{sample_size}
with shape parameter \code{shape} and scale parameter \code{scale}.
}
\description{
This function is able to generate random Pareto distributed data with 
the specified \code{shape} and \code{scale} parameters. The function 
has been written to be similar in type to the popular runif and rexp type 
of functions for generating data from a particular distribution.
}
\examples{
generate_pareto(10000, 5, 2)
generate_pareto(100, 15, 6)

}
